# backend/sdfcc-backend/src/main/java/de/quadflal/sdfccbackend/adapter/in/web/generated/api/RestaurantsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.quadflal.sdfccbackend.adapter.in.web.generated.api;

import de.quadflal.sdfccbackend.adapter.in.web.generated.model.AddRestaurantRequest;
import de.quadflal.sdfccbackend.adapter.in.web.generated.model.ErrorResponse;
import de.quadflal.sdfccbackend.adapter.in.web.generated.model.Restaurant;
import java.util.UUID;
import de.quadflal.sdfccbackend.adapter.in.web.generated.model.UpdateRestaurantRequest;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
public interface RestaurantsApi {

    String PATH_ADD_RESTAURANT = "/restaurants";
    /**
     * POST /restaurants : Add a new restaurant
     * Create a new restaurant in the database
     *
     * @param addRestaurantRequest  (required)
     * @return Restaurant created successfully (status code 201)
     *         or Invalid request body (status code 400)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = RestaurantsApi.PATH_ADD_RESTAURANT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Restaurant> addRestaurant(
         @Valid @RequestBody AddRestaurantRequest addRestaurantRequest
    );


    String PATH_DELETE_RESTAURANT_BY_ID = "/restaurants/{id}";
    /**
     * DELETE /restaurants/{id} : Delete a restaurant
     * Delete a restaurant by its ID
     *
     * @param id  (required)
     * @return Restaurant deleted successfully (status code 204)
     *         or Restaurant not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = RestaurantsApi.PATH_DELETE_RESTAURANT_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<Void> deleteRestaurantById(
         @PathVariable("id") UUID id
    );


    String PATH_GET_RESTAURANT_BY_ID = "/restaurants/{id}";
    /**
     * GET /restaurants/{id} : Get a restaurant by ID
     * Retrieve full restaurant details including address and location
     *
     * @param id  (required)
     * @return Restaurant found (status code 200)
     *         or Restaurant not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = RestaurantsApi.PATH_GET_RESTAURANT_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<Restaurant> getRestaurantById(
         @PathVariable("id") UUID id
    );


    String PATH_UPDATE_RESTAURANT = "/restaurants/{id}";
    /**
     * PUT /restaurants/{id} : Update a restaurant
     * Update an existing restaurant
     *
     * @param id  (required)
     * @param updateRestaurantRequest  (required)
     * @return Restaurant updated successfully (status code 200)
     *         or Invalid request body (status code 400)
     *         or Restaurant not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = RestaurantsApi.PATH_UPDATE_RESTAURANT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Restaurant> updateRestaurant(
         @PathVariable("id") UUID id,
         @Valid @RequestBody UpdateRestaurantRequest updateRestaurantRequest
    );

}